backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib==1.7.4
PyJWT==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
email-validator==2.1.0

